devpost/requirements.txt
google-cloud-translate==2.0.1
Scrapy==2.0.1
spacy_langdetect==0.1.2
spacy==2.2.4
pandas==1.0.3
